+++ tsMuxer/psgStreamReader.cpp
@@ -736,7 +736,7 @@ int PGSStreamReader::flushPacket(AVPacket& avPacket) { return 0; }
 
 void PGSStreamReader::setBuffer(uint8_t* data, int dataLen, bool lastBlock)
 {
-    if (m_tmpBufferLen + dataLen > m_tmpBuffer.size())
+    if ((size_t)(m_tmpBufferLen + dataLen) > m_tmpBuffer.size())
         m_tmpBuffer.resize(m_tmpBufferLen + dataLen);
 
     if (m_tmpBuffer.size() > 0)
