+++ tsMuxer/tsMuxer.cpp
@@ -420,7 +420,7 @@ bool TSMuxer::doFlush()
         newPCR = (m_endStreamDTS - m_minDts) / INT_FREQ_TO_TS_FREQ + 0.5 + m_fixed_pcr_offset;
         if (m_cbrBitrate != -1 && m_lastPCR != -1)
         {
-            int64_t cbrPCR = m_lastPCR + m_pcrBits * 90000.0 / m_cbrBitrate + 0.5;
+            uint64_t cbrPCR = m_lastPCR + m_pcrBits * 90000.0 / m_cbrBitrate + 0.5;
             newPCR = FFMAX(newPCR, cbrPCR);
         }
     }
@@ -507,7 +507,7 @@ bool TSMuxer::close()
 int TSMuxer::calcM2tsFrameCnt()
 {
     uint32_t byteCnt = 0;
-    for (int i = 0; i < m_m2tsDelayBlocks.size(); i++) byteCnt += m_m2tsDelayBlocks[i].second;
+    for (size_t i = 0; i < m_m2tsDelayBlocks.size(); i++) byteCnt += m_m2tsDelayBlocks[i].second;
     byteCnt -= m_prevM2TSPCROffset;
     byteCnt += m_outBufLen;
     assert(byteCnt % 192 == 0);
@@ -526,7 +526,7 @@ void TSMuxer::processM2TSPCR(int64_t pcrVal, int64_t pcrGAP)
     if (m_m2tsDelayBlocks.size() > 0)
     {
         int offset = m_prevM2TSPCROffset;
-        for (int i = 0; i < m_m2tsDelayBlocks.size(); i++)
+        for (size_t i = 0; i < m_m2tsDelayBlocks.size(); i++)
         {
             curPos = m_m2tsDelayBlocks[i].first + offset;
             int j = offset;
@@ -645,7 +645,7 @@ void TSMuxer::buildPesHeader(int pesStreamID, AVPacket& avPacket, int pid)
     int bufLen = pesPacket->getHeaderLength() + additionDataSize;
     m_pesData.resize(bufLen);
     memcpy(m_pesData.data(), tmpBuffer, bufLen);
-    for (int i = 0; i < tmpPriorityData.size(); ++i)
+    for (size_t i = 0; i < tmpPriorityData.size(); ++i)
         m_priorityData.push_back(
             std::pair<int, int>(tmpPriorityData[i].first + pesPacket->getHeaderLength(), tmpPriorityData[i].second));
 }
@@ -858,7 +858,7 @@ void TSMuxer::writePESPacket()
         uint8_t* curPtr = m_pesData.data();
         uint8_t* dataEnd = curPtr + m_pesData.size();
         bool payloadStart = true;
-        for (int i = 0; i < m_priorityData.size(); ++i)
+        for (size_t i = 0; i < m_priorityData.size(); ++i)
         {
             uint8_t* blockPtr = m_pesData.data() + m_priorityData[i].first;
             if (blockPtr > curPtr)
@@ -1369,7 +1369,7 @@ void TSMuxer::writeOutBuffer()
 void TSMuxer::parseMuxOpt(const std::string& opts)
 {
     vector<string> params = splitStr(opts.c_str(), ' ');
-    for (int i = 0; i < params.size(); i++)
+    for (size_t i = 0; i < params.size(); i++)
     {
         vector<string> paramPair = splitStr(trimStr(params[i]).c_str(), '=');
         if (paramPair.size() == 0)
@@ -1470,7 +1470,7 @@ void TSMuxer::openDstFile()
 vector<int64_t> TSMuxer::getFirstPts()
 {
     std::vector<int64_t> rez;
-    for (int i = 0; i < m_firstPts.size(); i++) rez.push_back(nanoClockToPts(m_firstPts[i]) + m_timeOffset);
+    for (size_t i = 0; i < m_firstPts.size(); i++) rez.push_back(nanoClockToPts(m_firstPts[i]) + m_timeOffset);
     return rez;
 }
 
@@ -1487,7 +1487,7 @@ void TSMuxer::alignPTS(TSMuxer* otherMuxer)
 vector<int64_t> TSMuxer::getLastPts()
 {
     std::vector<int64_t> rez;
-    for (int i = 0; i < m_lastPts.size(); i++) rez.push_back(nanoClockToPts(m_lastPts[i]) + m_timeOffset);
+    for (size_t i = 0; i < m_lastPts.size(); i++) rez.push_back(nanoClockToPts(m_lastPts[i]) + m_timeOffset);
     // if (!rez.empty())
     //    *rez.rbegin() += m_mainStreamFrameDuration;
     return rez;
@@ -1505,7 +1505,7 @@ void TSMuxer::setFileName(const std::string& fileName, FileFactory* fileFactory)
     m_fileNames.push_back(m_outFileName);
 }
 
-std::string TSMuxer::getFileNameByIdx(int idx)
+std::string TSMuxer::getFileNameByIdx(size_t idx)
 {
     if (idx < m_fileNames.size())
         return m_fileNames[idx];
